+++ src/commands/outputs/index.ts
@@ -3,6 +3,10 @@ export * from "./create-table.js";
 export * from "./delete-item.js";
 export * from "./delete-table.js";
 export * from "./get-item.js";
+export type {
+  ItemCollectionMetrics,
+  NativeItemCollectionMetrics,
+} from "./item-collection-metrics.js";
 export * from "./put-item.js";
 export * from "./query.js";
 export * from "./update-item.js";