+++ src/commands/outputs/put-item.ts
@@ -3,9 +3,20 @@ import { trusted } from "@infra-blocks/types";
 import type { Attributes } from "../../types.js";
 import { ifDefined, unsetUndefined } from "../lib.js";
 import { ConsumedCapacity } from "./consumed-capacity.js";
+import { ItemCollectionMetrics } from "./item-collection-metrics.js";
 
 export type PutItemOutput<T extends Attributes = Attributes> = {
+  /**
+   * The returned item, if any.
+   */
   item?: T;
+  /**
+   * The returned item collection metrics, if any.
+   */
+  itemCollectionMetrics?: ItemCollectionMetrics;
+  /**
+   * The returned consumed capacity, if any.
+   */
   consumedCapacity?: ConsumedCapacity;
 };
 
@@ -18,6 +29,10 @@ function decode<T extends Attributes = Attributes>(
 ): PutItemOutput<T> {
   return unsetUndefined({
     item: ifDefined(output.Attributes, trusted<T>),
+    itemCollectionMetrics: ifDefined(
+      output.ItemCollectionMetrics,
+      ItemCollectionMetrics.decode,
+    ),
     consumedCapacity: ifDefined(
       output.ConsumedCapacity,
       ConsumedCapacity.decode,