+++ test/unit/commands/outputs/index.ts
@@ -2,6 +2,7 @@ import { suite } from "node:test";
 import { consumedCapacityTests } from "./consumed-capacity.js";
 import { deleteItemTests } from "./delete-item.js";
 import { getItemTests } from "./get-item.js";
+import { itemCollectionMetricsTests } from "./item-collection-metrics.js";
 import { putItemTests } from "./put-item.js";
 import { queryTests } from "./query.js";
 import { updateItemTests } from "./update-item.js";
@@ -13,6 +14,7 @@ export const outputsTests = () => {
     deleteItemTests();
     getItemTests();
     queryTests();
+    itemCollectionMetricsTests();
     putItemTests();
     updateItemTests();
     writeTransactionTests();